+++ td/telegram/files/FileLoaderActor.h
@@ -17,7 +17,7 @@ class ResourceManager;
 class FileLoaderActor : public NetQueryCallback {
  public:
   virtual void set_resource_manager(ActorShared<ResourceManager>) = 0;
-  virtual void update_priority(int32 priority) = 0;
+  virtual void update_priority(int8 priority) = 0;
   virtual void update_resources(const ResourceState &other) = 0;
 
   // TODO: existence of this function is a dirty hack. Refactoring is highly appreciated
